Introduce `IMicrosoftIdentityAuthenticationDelegatingHandlerFactory`
Currently, it's very difficult to customize the behavior of the
`DelegatingHandler` classes that add the auth token, as the
`DownstreamWebApiExtensions` `new()` up the handlers directly.

Ideally I would register the handlers in the `IServiceCollection`
directly, and then resolve it from the container, however in the case of
multiple downstream APIs we need to resolve the correct named
configuration from the `IOptionsMonitor`, and I'm not aware of a way to
do that with just the dependency injection abstraction APIs.

As a result, I've instead introduced a public factory interface that
takes the `IServiceProvider` and the configuration name, which then
allows for creating the handler from the factory class. Users that want
to customize the handlers can then replace the
`IMicrosoftIdentityAuthenticationDelegatingHandlerFactory` registration
with a class that creates their own implementations.