Refactor to remove `AllowWebApiToBeAuthorizedByACL` check from callback
Refactor to move the `AllowWebApiToBeAuthorizedByACL` check to before
the `OnTokenValidated` event chain. This is both a minor performance win
in the ACL case (possibly removing an event per token if the user
doesn't already have an event), and is a minor simplification of the
conditional logic.

As a result it's no longer easy to validate the behavior in the ACL
case; if that's deemed important we can move the whole logic set into
another helper method.