Switch `UnauthorizedAccessException` to `TokenValidatedContext.Fail`
Today, if a token is validated, but is:

1. Missing both Scopes and Roles
2. Not authorized via ACL

Then the `JwtBearerHandler` will throw an `UnauthorizedAccessException`
that looks like this:

System.UnauthorizedAccessException: IDW10201: Neither scope or roles claim was found in the bearer token.
  at Microsoft.Identity.Web.MicrosoftIdentityWebApiAuthenticationBuilderExtensions.<>c__DisplayClass3_1.<<AddMicrosoftIdentityWebApiImplementation>b__1>d.MoveNext()
  at Microsoft.AspNetCore.Authentication.JwtBearer.JwtBearerHandler.HandleAuthenticateAsync()
  at Microsoft.AspNetCore.Authentication.JwtBearer.JwtBearerHandler.HandleAuthenticateAsync()
  at Microsoft.AspNetCore.Authentication.AuthenticationHandler`1.AuthenticateAsync()
  at Microsoft.AspNetCore.Authentication.AuthenticationService.AuthenticateAsync(HttpContext context, String scheme)
  at Microsoft.AspNetCore.Authentication.AuthenticationMiddleware.Invoke(HttpContext context)

which by default results in an unhandled exception in the auth
middleware pipeline, which in turn results in a 500 error for the user.

While one could make the argument that a 500 is appropriate in this
scenario, as the most likely cause is a misconfiguration on the AAD
side, similar types of misconfiguration, such as a missing `tid` or
`tenantId` claim results in a 401 Unauthorized error.

As a result, replacing the throw statement with a call to mark the
`ResultContext` as failed.