Fix for #2371
In .NET 5+, we now use the DefaultTokenAcquisitionHost (the host for SDK apps) instead of the
Asp.NET core one, when the service collection was not initialized by ASP.NET Core (that is
the `IWebHostEnvironment` is not present in the collection.

If developers want the ASP.NET Core host, they would need to use the WebApplication.CreateBuilder().Services instead
of instanciating a simple service collection.