Update OwinWebApp DevApp to configure the redirect_uri parameter.
Current implementation seems to suggest that this parameter is configured
in the Startup.Auth.cs file, as part of the ConfidentialClientApplicationOptions
however this is not realted with the redirect_uri parameter used by the
Identity Provider to redirect back to the relying party after the authentication
is completed. This parameter is actually configured in OpenIdConnectAuthenticationOptions
parameter.