Add checks to protect the internal claims used by MIW. Ref: issue #2968 (#3131)
* Add InternalClaimDetectedException for ID Token validation

Introduce InternalClaimDetectedException in Microsoft.Identity.Web and Microsoft.Identity.Web.OWIN namespaces. This exception is thrown when internal ID Token claims (UniqueTenantIdentifier, UniqueObjectIdentifier) are detected in the user's ID Token. Updated AppBuilderExtension.cs and MicrosoftIdentityWebAppAuthenticationBuilder.cs to check for these claims and throw the exception if found. The new exception class includes a property to store the invalid claim

* Add unit tests for the modifications

* Add check for claim equality

* Change string comparison to OrdinalIgnoreCase

* Improvements based on feedback from PR

* PR changes post discussion

* Update src/Microsoft.Identity.Web.OWIN/AppBuilderExtension.cs

Update the description based on PR review

Co-authored-by: Bogdan Gavril <bogavril@microsoft.com>

* Refactor to test each claim separately.

* Change the logic to match

---------

Co-authored-by: Bogdan Gavril <bogavril@microsoft.com>